requirements.txt
PyGithub
yaml